refactor: move go files to the go folder
