+++ .github/workflows/go-test.yaml
@@ -11,8 +11,7 @@ on:
     branches:
       - main
     paths:
-      - "lib/**"
-      - "*.go"
+      - "go/**"
       - "go.mod"
       - "go.sum"
   workflow_dispatch:
@@ -24,7 +23,7 @@ jobs:
   test:
     strategy:
       matrix:
-        os: [macos-latest, ubuntu-latest, windows-latest]
+        os: [macos-12, macos-latest, ubuntu-latest, windows-latest]
     runs-on: ${{ matrix.os }}
     steps:
       - name: Git checkout
@@ -32,6 +31,6 @@ jobs:
       - name: Set up Go
         uses: actions/setup-go@v2
         with:
-          go-version: 1.21
+          go-version: 1.22
       - name: Go code test
         run: go test ./...
