+++ go/lib/darwin-amd64/dummy.go
