+++ go/lib/dummy.go
@@ -1,0 +1,10 @@
+package lib
+
+import (
+	_ "kcl-lang.io/lib/go/lib/darwin-amd64"
+	_ "kcl-lang.io/lib/go/lib/darwin-arm64"
+	_ "kcl-lang.io/lib/go/lib/linux-amd64"
+	_ "kcl-lang.io/lib/go/lib/linux-arm64"
+	_ "kcl-lang.io/lib/go/lib/windows-amd64"
+	_ "kcl-lang.io/lib/go/lib/windows-arm64"
+)
