+++ go/lib/kclvm_cli_shared_lib_linux_arm64.go
@@ -1,0 +1,11 @@
+package lib
+
+import (
+	_ "embed"
+)
+
+//go:embed linux-arm64/libkclvm_cli_cdylib.so
+var CliLib []byte
+
+//go:embed linux-arm64/kclvm_cli
+var CliBin []byte
