+++ go/lib/kclvm_cli_shared_lib_windows_amd64.go
@@ -1,0 +1,14 @@
+package lib
+
+import (
+	_ "embed"
+)
+
+//go:embed windows-amd64/kclvm_cli_cdylib.dll
+var CliLib []byte
+
+//go:embed windows-amd64/kclvm_cli_cdylib.lib
+var ExportLib []byte
+
+//go:embed windows-amd64/kclvm_cli.exe
+var CliBin []byte
