+++ go/lib/linux-amd64/dummy.go
