+++ go/lib/windows-amd64/dummy.go
