--- kclvm_cli_shared_lib_darwin_arm64.go
@@ -1,11 +1,0 @@
-package lib
-
-import (
-	_ "embed"
-)
-
-//go:embed lib/darwin-arm64/libkclvm_cli_cdylib.dylib
-var kclvmCliLib []byte
-
-//go:embed lib/darwin-arm64/kclvm_cli
-var kclvmCliBin []byte
