--- kclvm_cli_shared_lib_linux_amd64.go
@@ -1,11 +1,0 @@
-package lib
-
-import (
-	_ "embed"
-)
-
-//go:embed lib/linux-amd64/libkclvm_cli_cdylib.so
-var kclvmCliLib []byte
-
-//go:embed lib/linux-amd64/kclvm_cli
-var kclvmCliBin []byte
