--- kclvm_cli_shared_lib_linux_arm64.go
@@ -1,11 +1,0 @@
-package lib
-
-import (
-	_ "embed"
-)
-
-//go:embed lib/linux-arm64/libkclvm_cli_cdylib.so
-var kclvmCliLib []byte
-
-//go:embed lib/linux-arm64/kclvm_cli
-var kclvmCliBin []byte
