--- lib/dummy.go
@@ -1,10 +1,0 @@
-package lib
-
-import (
-	_ "kcl-lang.io/lib/lib/darwin-amd64"
-	_ "kcl-lang.io/lib/lib/darwin-arm64"
-	_ "kcl-lang.io/lib/lib/linux-amd64"
-	_ "kcl-lang.io/lib/lib/linux-arm64"
-	_ "kcl-lang.io/lib/lib/windows-amd64"
-	_ "kcl-lang.io/lib/lib/windows-arm64"
-)
