+++ spec/gpyrpc/Makefile
@@ -3,6 +3,6 @@ run:
 	protoc spec.proto --kotlin_out ../../kotlin/src/main/kotlin
 	protoc spec.proto --pyi_out ../../python/kcl_lib/api
 	protoc spec.proto --python_out ../../python/kcl_lib/api
-	protoc spec.proto --go_out ../../
+	protoc spec.proto --go_out ../../go/api
 	protoc spec.proto --csharp_out ../../dotnet/KclLib/api
 	protoc spec.proto --cpp_out ../../cpp/src/api
